Loan-Chat-bot/requirements.txt
openai
pandas
numpy
faiss-cpu
gradio
sentence-transformers
python-dotenv